requirements.txt
pandas>=2.2.2
SQLAlchemy>=2.0.25
pyodbc>=5.0.1
numpy>=1.26.4
python-dotenv>=1.0.1